requirements.txt
openai
groq
python-dotenv

google-play-scraper
pandas
requests
beautifulsoup4
ipython

nltk
num2words
